pyproject.toml
[tool.poetry]
name = "meme-bot"
version = "0.1.0"
description = ""
authors = ["Nimrod Oren <nimrod.oren@zesty.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
pygithub = "^2.3.0"
jwt = "^1.3.1"
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
sqlmodel = "^0.0.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
